requirements.txt
requests
sorl-thumbnail
django
asgiref==3.2.10
mccabe==0.6.1
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
sqlparse==0.3.1